Reorganize vim configuration
Previously, the entire configuration was in .vimrc. With inspiration from @ChrisToomey I split it up into ftplugins, plugin_config and other_config. Mappings are helpfully provided to facilitate quick lookup of any setting.

vim-enmasse didn't want to work with that setting, so for now it was added directly in the vimrc. I have no idea why that is.