Add workaround for promptline
Why was the change made?

The fonts for promptline didn't work in Alacritty. I had very little
time to fix it so I removed the symbols. While doing that I looked up
the setting to remove the part of the prompt that contains the user
name, I don't feel I need that.