Set .tmux filetype to sh
**Why** is the change needed?

The commands in the .tmux files are run in the shell, talking to tmux.
It thus makes the most sense to treat them as such.

**How** is the need addressed?

Add an autocommand to change_filetype that maps the filetype of any file
named `.tmux` to `sh`.

What **side-effects** could the change have?

If any other files are name `.tmux`, it changes their filetype as well.