Add month to status bar
**Why** is the change needed?

I want to tmux to be self contained (like my operating system) and
I should be able to see the date at once, since it is useful for many
planning-related activities.

**How** is the need addressed?

Use strftime to add the date to the right side of the status bar.

What **side-effects** could the change have?

It is currently in Dutch, would prefer Swedish or English. It doesn't
really matter.