Organize paths.sh according to standard
**Why** is the change needed?

The standard Ubuntu PATH for interactive usage, in most cases is:
PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin"

The above is preffered hardcoded PATH for kernel, libc, shells, and similar use cases. For example, this is the built-in path
that should be used by /bin/sh, if PATH is not already set in the environment.

https://wiki.ubuntu.com/PATH