Turn off hardtime by default
**Why** is the change needed?

I can turn it on if I want to practice in a given session, but it's nice
that it's off in all the one-off vim instances I launch throughout the
day.

It's quite possible I'll turn it on again in the future if I feel like
I'm getting lazy with the economy of movement.